+++ src/renderer/src/components/ui/modal/stacked/atom.ts
@@ -1,0 +1,5 @@
+import { atom } from "jotai"
+
+import type { ModalProps } from "./types"
+
+export const modalStackAtom = atom([] as (ModalProps & { id: string })[])
