+++ src/renderer/src/components/ui/modal/stacked/context.tsx
@@ -1,10 +1,5 @@
-import { atom } from "jotai"
 import type { FC, RefObject } from "react"
-import { createContext, useContext } from "react"
-
-import type { ModalProps } from "./types"
-
-export const modalIdToPropsMap = {} as Record<string, ModalProps>
+import { createContext } from "react"
 
 export type CurrentModalContentProps = ModalContentPropsInternal & {
   ref: RefObject<HTMLElement | null>
@@ -14,10 +9,7 @@ export const CurrentModalContext = createContext<CurrentModalContentProps>(
   null as any,
 )
 
-export const useCurrentModal = () => useContext(CurrentModalContext)
-
 export type ModalContentComponent<T> = FC<ModalContentPropsInternal & T>
 export type ModalContentPropsInternal = {
   dismiss: () => void
 }
-export const modalStackAtom = atom([] as (ModalProps & { id: string })[])
