+++ src/renderer/src/components/ui/modal/stacked/index.ts
@@ -1,5 +1,6 @@
 export * from "./context"
 export * from "./helper"
+export * from "./hooks"
 export * from "./modal"
 export * from "./provider"
 export * from "./types"
